@@ -1,55 +1,32 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-	<Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
+	  <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <PackageId>Greg</PackageId>
     <Authors></Authors>
-    <Company>Autodesk</Company>
-    <ProductVersion>8.0.30703</ProductVersion>
-    <SchemaVersion>2.0</SchemaVersion>
+    <Company>Autodesk, Inc</Company>
+    <AssemblyCopyright>Copyright � Autodesk, Inc 2022</AssemblyCopyright>
     <ProjectGuid>{644207B4-7E7F-474A-952E-3453960D8A01}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>Greg</RootNamespace>
     <AssemblyName>Greg</AssemblyName>
-    <TargetFrameworks>net452;net48;net5.0;net6.0</TargetFrameworks>
-    <FileAlignment>512</FileAlignment>
-    <TargetFrameworkProfile />
-	  <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
-	  <Deterministic>false</Deterministic>
-	  <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-    <GenerateNuspecDependsOn>$(GenerateNuspecDependsOn);ReadPackageVersionFromOutputAssembly</GenerateNuspecDependsOn>
-	  <PackageIcon>logo_square_32x32.png</PackageIcon>
-	  <Description>The Dynamo Package Manager .Net Client.</Description>
+    <TargetFrameworks>net8.0</TargetFrameworks>
+    <TotalDays>$([System.DateTime]::op_Subtraction($([System.DateTime]::get_Now().get_Date()),$([System.DateTime]::new(2023,1,1))).get_TotalDays())</TotalDays>
+    <Days>$([MSBuild]::Multiply($(TotalDays), 10))</Days>
+    <Hours>$([MSBuild]::Divide($([System.DateTime]::get_Now().get_TimeOfDay().get_TotalHours()), 3))</Hours>
+    <Revision>$([MSBuild]::Add($(Days), $(Hours)).ToString('F0'))</Revision>
+    <Version>3.0.2.$(Revision)</Version>
+    <GeneratedAssemblyInfoFile>$(MSBuildProjectDirectory)/$(AppDesignerFolder)/AssemblyVersionInfo.cs</GeneratedAssemblyInfoFile>
+    <GenerateAssemblyInfo>true</GenerateAssemblyInfo>
+    <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+    <PackageIcon>logo_square_32x32.png</PackageIcon>
+    <Description>The Dynamo Package Manager .Net Client.</Description>
+    <OutputPath>..\..\bin\$(Configuration)\</OutputPath>
   </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>..\..\bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <Prefer32Bit>false</Prefer32Bit>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>pdbonly</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>..\..\bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <Prefer32Bit>false</Prefer32Bit>
-  </PropertyGroup>
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net452' OR '$(TargetFramework)' == 'net48'">
-    <Reference Include="System.Configuration" />
-    <Reference Include="System.IO.Compression" />
-    <Reference Include="System.IO.Compression.FileSystem" />
-    <Reference Include="System.Web" />
-  </ItemGroup>
+
   <ItemGroup Label="Common package dependencies">
-    <PackageReference Include="RestSharp" Version="106.12.0" />
-	<PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+    <PackageReference Include="RestSharp" Version="108.0.1" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\..\CHANGELOG.md">
